Move reset_fragments_dir to tests.helpers.changelog
Importing from a conftest module is discouraged by pytest. The helper
belongs in tests/helpers, alongside the other shared test utilities.
The repo_with_towncrier fixture stays in cli/release/conftest.py since
fixtures are auto-discovered, not imported.
